requirements.txt
Flask==2.3.3
openpyxl==3.1.2
reportlab==4.0.4
Werkzeug==2.3.7
matplotlib==3.7.2

# 云端部署依赖
gunicorn==21.2.0
psycopg2-binary==2.9.9
